+++ src/bunit.core/ComponentParameterFactory.cs
@@ -6,110 +6,99 @@ using Bunit.Rendering;
 using Microsoft.AspNetCore.Components;
 using Microsoft.AspNetCore.Components.Rendering;
 
-using EC = Microsoft.AspNetCore.Components.EventCallback;
-
 namespace Bunit
 {
 	/// <summary>
-	/// Base class for test classes that contains Razor component tests.
+	/// <see cref="ComponentParameter"/> factory methods.
 	/// </summary>
-	public abstract class ComponentTestFixture : TestContext
+	public static class ComponentParameterFactory
 	{
 		/// <summary>
-		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> as parameter value 
-		/// for this <see cref="TestContext"/> and
-		/// <paramref name="callback"/>.
+		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> that will call the provided <paramref name="callback"/>.
 		/// </summary>
 		/// <param name="name">Parameter name.</param>
 		/// <param name="callback">The event callback.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected ComponentParameter EventCallback(string name, Action callback)
+		public static ComponentParameter EventCallback(string name, Action callback)
 		{
-			return ComponentParameter.CreateParameter(name, EC.Factory.Create(this, callback));
+			return ComponentParameter.CreateParameter(name, new EventCallback(null, callback));
 		}
 
 		/// <summary>
-		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> as parameter value for this <see cref="TestContext"/> and
-		/// <paramref name="callback"/>.
+		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> that will call the provided <paramref name="callback"/>.
 		/// </summary>
 		/// <param name="name">Parameter name.</param>
 		/// <param name="callback">The event callback.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected ComponentParameter EventCallback(string name, Action<object> callback)
+		public static ComponentParameter EventCallback(string name, Action<object> callback)
 		{
-			return ComponentParameter.CreateParameter(name, EC.Factory.Create(this, callback));
+			return ComponentParameter.CreateParameter(name, new EventCallback(null, callback));
 		}
 
 		/// <summary>
-		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> as parameter value for this <see cref="TestContext"/> and
-		/// <paramref name="callback"/>.
+		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> that will call the provided <paramref name="callback"/>.
 		/// </summary>
 		/// <param name="name">Parameter name.</param>
 		/// <param name="callback">The event callback.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected ComponentParameter EventCallback(string name, Func<Task> callback)
+		public static ComponentParameter EventCallback(string name, Func<Task> callback)
 		{
-			return ComponentParameter.CreateParameter(name, EC.Factory.Create(this, callback));
+			return ComponentParameter.CreateParameter(name, new EventCallback(null, callback));
 		}
 
 		/// <summary>
-		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> as parameter value for this <see cref="TestContext"/> and
-		/// <paramref name="callback"/>.
+		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> that will call the provided <paramref name="callback"/>.
 		/// </summary>
 		/// <param name="name">Parameter name.</param>
 		/// <param name="callback">The event callback.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected ComponentParameter EventCallback(string name, Func<object, Task> callback)
+		public static ComponentParameter EventCallback(string name, Func<object, Task> callback)
 		{
-			return ComponentParameter.CreateParameter(name, EC.Factory.Create(this, callback));
+			return ComponentParameter.CreateParameter(name, new EventCallback(null, callback));
 		}
 
 		/// <summary>
-		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> as parameter value for this <see cref="TestContext"/> and
-		/// <paramref name="callback"/>.
+		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> that will call the provided <paramref name="callback"/>.
 		/// </summary>
 		/// <param name="name">Parameter name.</param>
 		/// <param name="callback">The event callback.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected ComponentParameter EventCallback<TValue>(string name, Action callback)
+		public static ComponentParameter EventCallback<TValue>(string name, Action callback)
 		{
-			return ComponentParameter.CreateParameter(name, EC.Factory.Create<TValue>(this, callback));
+			return ComponentParameter.CreateParameter(name, new EventCallback<TValue>(null, callback));
 		}
 
 		/// <summary>
-		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> as parameter value for this <see cref="TestContext"/> and
-		/// <paramref name="callback"/>.
+		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> that will call the provided <paramref name="callback"/>.
 		/// </summary>
 		/// <param name="name">Parameter name.</param>
 		/// <param name="callback">The event callback.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected ComponentParameter EventCallback<TValue>(string name, Action<TValue> callback)
+		public static ComponentParameter EventCallback<TValue>(string name, Action<TValue> callback)
 		{
-			return ComponentParameter.CreateParameter(name, EC.Factory.Create<TValue>(this, callback));
+			return ComponentParameter.CreateParameter(name, new EventCallback<TValue>(null, callback));
 		}
 
 		/// <summary>
-		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> as parameter value for this <see cref="TestContext"/> and
-		/// <paramref name="callback"/>.
+		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> that will call the provided <paramref name="callback"/>.
 		/// </summary>
 		/// <param name="name">Parameter name.</param>
 		/// <param name="callback">The event callback.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected ComponentParameter EventCallback<TValue>(string name, Func<Task> callback)
+		public static ComponentParameter EventCallback<TValue>(string name, Func<Task> callback)
 		{
-			return ComponentParameter.CreateParameter(name, EC.Factory.Create<TValue>(this, callback));
+			return ComponentParameter.CreateParameter(name, new EventCallback<TValue>(null, callback));
 		}
 
 		/// <summary>
-		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> as parameter value for this <see cref="TestContext"/> and
-		/// <paramref name="callback"/>.
+		/// Creates a <see cref="ComponentParameter"/> with an <see cref="Microsoft.AspNetCore.Components.EventCallback"/> that will call the provided <paramref name="callback"/>.
 		/// </summary>
 		/// <param name="name">Parameter name.</param>
 		/// <param name="callback">The event callback.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected ComponentParameter EventCallback<TValue>(string name, Func<TValue, Task> callback)
+		public static ComponentParameter EventCallback<TValue>(string name, Func<TValue, Task> callback)
 		{
-			return ComponentParameter.CreateParameter(name, EC.Factory.Create<TValue>(this, callback));
+			return ComponentParameter.CreateParameter(name, new EventCallback<TValue>(null, callback));
 		}
 
 		/// <summary>
@@ -118,7 +107,7 @@ namespace Bunit
 		/// <param name="name">Parameter name</param>
 		/// <param name="value">Value or null of the parameter</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter Parameter(string name, object? value)
+		public static ComponentParameter Parameter(string name, object? value)
 		{
 			return ComponentParameter.CreateParameter(name, value);
 		}
@@ -129,7 +118,7 @@ namespace Bunit
 		/// <param name="name">Parameter name</param>
 		/// <param name="value">Value of the parameter</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter CascadingValue(string name, object value)
+		public static ComponentParameter CascadingValue(string name, object value)
 		{
 			return ComponentParameter.CreateCascadingValue(name, value);
 		}
@@ -139,7 +128,7 @@ namespace Bunit
 		/// </summary>
 		/// <param name="value">Value of the parameter</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter CascadingValue(object value)
+		public static ComponentParameter CascadingValue(object value)
 		{
 			return ComponentParameter.CreateCascadingValue(null, value);
 		}
@@ -150,7 +139,7 @@ namespace Bunit
 		/// </summary>
 		/// <param name="markup">Markup to pass to the child content parameter</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter ChildContent(string markup)
+		public static ComponentParameter ChildContent(string markup)
 		{
 			return RenderFragment(nameof(ChildContent), markup);
 		}
@@ -162,7 +151,7 @@ namespace Bunit
 		/// <typeparam name="TComponent">The type of the component to render with the <see cref="Microsoft.AspNetCore.Components.RenderFragment"/></typeparam>
 		/// <param name="parameters">Parameters to pass to the <typeparamref name="TComponent"/>.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter ChildContent<TComponent>(params ComponentParameter[] parameters) where TComponent : class, IComponent
+		public static ComponentParameter ChildContent<TComponent>(params ComponentParameter[] parameters) where TComponent : class, IComponent
 		{
 			return RenderFragment<TComponent>(nameof(ChildContent), parameters);
 		}
@@ -174,7 +163,7 @@ namespace Bunit
 		/// <param name="name">Parameter name.</param>
 		/// <param name="markup">Markup to pass to the render fragment parameter</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter RenderFragment(string name, string markup)
+		public static ComponentParameter RenderFragment(string name, string markup)
 		{
 			return ComponentParameter.CreateParameter(name, markup.ToMarkupRenderFragment());
 		}
@@ -187,7 +176,7 @@ namespace Bunit
 		/// <param name="name">Parameter name.</param>
 		/// <param name="parameters">Parameters to pass to the <typeparamref name="TComponent"/>.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter RenderFragment<TComponent>(string name, params ComponentParameter[] parameters) where TComponent : class, IComponent
+		public static ComponentParameter RenderFragment<TComponent>(string name, params ComponentParameter[] parameters) where TComponent : class, IComponent
 		{
 			return ComponentParameter.CreateParameter(name, parameters.ToComponentRenderFragment<TComponent>());
 		}
@@ -200,7 +189,7 @@ namespace Bunit
 		/// <param name="name">Parameter name.</param>
 		/// <param name="template"><see cref="Microsoft.AspNetCore.Components.RenderFragment{TValue}" /> to pass to the parameter.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter Template<TValue>(string name, RenderFragment<TValue> template)
+		public static ComponentParameter Template<TValue>(string name, RenderFragment<TValue> template)
 		{
 			return ComponentParameter.CreateParameter(name, template);
 		}
@@ -214,7 +203,7 @@ namespace Bunit
 		/// <param name="name">Parameter name.</param>
 		/// <param name="markupFactory">A markup factory that takes a <typeparamref name="TValue"/> as input and returns markup/HTML.</param>
 		/// <returns>The <see cref="ComponentParameter"/>.</returns>
-		protected static ComponentParameter Template<TValue>(string name, Func<TValue, string> markupFactory)
+		public static ComponentParameter Template<TValue>(string name, Func<TValue, string> markupFactory)
 		{
 			return Template<TValue>(name, value => (RenderTreeBuilder builder) => builder.AddMarkupContent(0, markupFactory(value)));
 		}