+++ src/bunit.core/IRenderedComponentBase.cs
@@ -1,3 +1,4 @@
+using System;
 using Bunit.Rendering;
 
 using Microsoft.AspNetCore.Components;
@@ -8,7 +9,7 @@ namespace Bunit
 	/// Represents a rendered component-under-test. 
 	/// </summary>
 	/// <typeparam name="TComponent">The type of the component under test</typeparam>
-	public interface IRenderedComponentBase<out TComponent> : IRenderedFragmentBase where TComponent : IComponent
+	public interface IRenderedComponentBase<TComponent> : IRenderedFragmentBase where TComponent : IComponent
 	{
 		/// <summary>
 		/// Gets the component under test
@@ -21,15 +22,21 @@ namespace Bunit
 		void Render();
 
 		/// <summary>
-		/// Render the component under test again.
+		/// Render the component under test again with the provided <paramref name="parameters"/>.
 		/// </summary>
 		/// <param name="parameters">Parameters to pass to the component upon rendered</param>
 		void SetParametersAndRender(ParameterView parameters);
 
 		/// <summary>
-		/// Render the component under test again.
+		/// Render the component under test again with the provided <paramref name="parameters"/>.
 		/// </summary>
 		/// <param name="parameters">Parameters to pass to the component upon rendered</param>
 		void SetParametersAndRender(params ComponentParameter[] parameters);
+
+		/// <summary>
+		/// Render the component under test again with the provided parameters from the <paramref name="parameterBuilder"/>.
+		/// </summary>
+		/// <param name="parameterBuilder">An action that receives a <see cref="ComponentParameterBuilder{TComponent}"/>.</param>
+		void SetParametersAndRender(Action<ComponentParameterBuilder<TComponent>> parameterBuilder);
 	}
 }