+++ src/bunit.core/Rendering/ComponentParameter.cs
@@ -25,7 +25,13 @@ namespace Bunit.Rendering
 		/// </summary>
 		public bool IsCascadingValue { get; }
 
-		private ComponentParameter(string? name, object? value, bool isCascadingValue)
+		/// <summary>
+		/// Creates an instance of the <see cref="ComponentParameter"/> type.
+		/// </summary>
+		/// <param name="name">An optional name</param>
+		/// <param name="value">An optional value</param>
+		/// <param name="isCascadingValue">Whether or not this is a cascading value</param>
+		internal ComponentParameter(string? name, object? value, bool isCascadingValue)
 		{
 			if (isCascadingValue && value is null)
 				throw new ArgumentNullException(nameof(value), "Cascading values cannot be set to null");