+++ src/bunit.web.tests/RefreshableQueryCollectionTest.cs
@@ -1,4 +1,4 @@
-﻿using Bunit.TestAssets.SampleComponents;
+using Bunit.TestAssets.SampleComponents;
 
 using Shouldly;
 
@@ -6,7 +6,7 @@ using Xunit;
 
 namespace Bunit
 {
-	public class RefreshableQueryCollectionTest : ComponentTestFixture
+	public class RefreshableQueryCollectionTest : TestContext
 	{
 		[Fact(DisplayName = "When the query returns no elements, the collection is empty")]
 		public void Test001()