+++ src/bunit.web.tests/Rendering/Internal/HtmlizerTests.cs
@@ -1,68 +1,32 @@
+using Microsoft.AspNetCore.Components;
+using Microsoft.AspNetCore.Components.Rendering;
+using Microsoft.AspNetCore.Components.Web;
+using Shouldly;
+using Xunit;
+
 namespace Bunit.Rendering.Internal
 {
-	using Microsoft.AspNetCore.Components;
-	using Microsoft.AspNetCore.Components.Rendering;
-	using Microsoft.AspNetCore.Components.Web;
-
-	using Xunit;
-
-	public class HtmlizerTests : ComponentTestFixture
+	public class HtmlizerTests : TestContext
 	{
-		public static TheoryData GetOnClickData()
-		{
-			return new TheoryData<bool, bool>
-			{
-				{ false, true },
-				{ true, false },
-				{ true, true }
-			};
-		}
-
-		[Theory(DisplayName =
-			"The component is rendered without internal Blazor attributes, " +
-			"regardless of the PreventDefault and StopPropagation settings.")]
-		[MemberData(nameof(GetOnClickData))]
-		public void Test001(bool stopPropagation, bool preventDefault)
-		{
-			//Arrange
-			var component = RenderComponent<Htmlizer01Component>(
-				EventCallback<MouseEventArgs>(nameof(Htmlizer01Component.OnClick), OnClickCallback),
-				Parameter(nameof(Htmlizer01Component.OnClickStopPropagation), stopPropagation),
-				Parameter(nameof(Htmlizer01Component.OnClickPreventDefault), preventDefault));
-
-			//Act
-
-			//Assert
-			component.MarkupMatches("<button type=\"button\">Click me!</button>");
-			void OnClickCallback(MouseEventArgs e)
-			{
-				// NOTE: This line is only for the completeness of the EventCallback
-				Assert.NotNull(e);
-			}
-		}
-
 		[Theory(DisplayName = "The component contains correctly prefixed internal attributes.")]
-		[MemberData(nameof(GetOnClickData))]
+		[InlineData(false, true)]
+		[InlineData(true, false)]
+		[InlineData(true, true)]
 		public void Test002(bool stopPropagation, bool preventDefault)
 		{
 			//Arrange
-			var component = RenderComponent<Htmlizer01Component>(
-				EventCallback<MouseEventArgs>(nameof(Htmlizer01Component.OnClick), OnClickCallback),
-				Parameter(nameof(Htmlizer01Component.OnClickStopPropagation), stopPropagation),
-				Parameter(nameof(Htmlizer01Component.OnClickPreventDefault), preventDefault));
+			var component = RenderComponent<Htmlizer01Component>(parameters => parameters
+				.Add(p => p.OnClick, (MouseEventArgs e) => { })
+				.Add(p => p.OnClickStopPropagation, stopPropagation)
+				.Add(p => p.OnClickPreventDefault, preventDefault)
+			);
 
 			//Act
 			var button = component.Find("button");
 
 			//Assert
-			Assert.Equal(stopPropagation, button.HasAttribute("blazor:__internal_stopPropagation_onclick"));
-			Assert.Equal(preventDefault, button.HasAttribute("blazor:__internal_preventDefault_onclick"));
-
-			void OnClickCallback(MouseEventArgs e)
-			{
-				// NOTE: This line is only for the completeness of the EventCallback
-				Assert.NotNull(e);
-			}
+			button.HasAttribute(Htmlizer.ToBlazorAttribute("__internal_stopPropagation_onclick")).ShouldBe(stopPropagation);
+			button.HasAttribute(Htmlizer.ToBlazorAttribute("__internal_preventDefault_onclick")).ShouldBe(preventDefault);
 		}
 
 		private class Htmlizer01Component : ComponentBase
@@ -76,7 +40,6 @@ namespace Bunit.Rendering.Internal
 			[Parameter]
 			public bool OnClickStopPropagation { get; set; }
 
-			/// <inheritdoc />
 			protected override void BuildRenderTree(RenderTreeBuilder builder)
 			{
 				base.BuildRenderTree(builder);
@@ -86,7 +49,7 @@ namespace Bunit.Rendering.Internal
 
 				if (OnClick.HasDelegate)
 				{
-					builder.AddAttribute(2, "onclick", Microsoft.AspNetCore.Components.EventCallback.Factory.Create(this, OnClick));
+					builder.AddAttribute(2, "onclick", EventCallback.Factory.Create(this, OnClick));
 					builder.AddEventStopPropagationAttribute(3, "onclick", OnClickStopPropagation);
 					builder.AddEventPreventDefaultAttribute(4, "onclick", OnClickPreventDefault);
 				}