+++ src/bunit.web/Rendering/RenderedComponent.cs
@@ -1,6 +1,6 @@
 using System;
 using System.Collections.Generic;
-
+using System.Linq;
 using Microsoft.AspNetCore.Components;
 
 namespace Bunit.Rendering
@@ -20,31 +20,47 @@ namespace Bunit.Rendering
 		public void Render() => SetParametersAndRender(ParameterView.Empty);
 
 		/// <inheritdoc/>
-		public void SetParametersAndRender(params ComponentParameter[] parameters)
-		{
-			var parameterView = ParameterView.Empty;
-			if (parameters.Length > 0)
-			{
-				var paramDict = new Dictionary<string, object?>(parameters.Length);
-				foreach (var param in parameters)
-				{
-					if (param.IsCascadingValue)
-						throw new InvalidOperationException($"You cannot provide a new cascading value through the {nameof(SetParametersAndRender)} method.");
-
-					paramDict.Add(param.Name!, param.Value);
-				}
-				parameterView = ParameterView.FromDictionary(paramDict);
-			}
-			SetParametersAndRender(parameterView);
-		}
-
-		/// <inheritdoc/>
 		public void SetParametersAndRender(ParameterView parameters)
 		{
 			Renderer.InvokeAsync(() =>
 			{
 				Instance.SetParametersAsync(parameters);
 			});
+		}
+
+		/// <inheritdoc/>
+		public void SetParametersAndRender(params ComponentParameter[] parameters)
+		{
+			SetParametersAndRender(ToParameterView(parameters));
+		}
+
+		/// <inheritdoc/>
+		public void SetParametersAndRender(Action<ComponentParameterBuilder<TComponent>> parameterBuilder)
+		{
+			if (parameterBuilder is null)
+				throw new ArgumentNullException(nameof(parameterBuilder));
+
+			var builder = new ComponentParameterBuilder<TComponent>();
+			parameterBuilder(builder);
+
+			SetParametersAndRender(ToParameterView(builder.Build()));
+		}
+
+		private static ParameterView ToParameterView(IReadOnlyList<ComponentParameter> parameters)
+		{
+			var parameterView = ParameterView.Empty;
+			if (parameters.Any())
+			{
+				var paramDict = new Dictionary<string, object?>();
+				foreach (var param in parameters)
+				{
+					if (param.IsCascadingValue)
+						throw new InvalidOperationException($"You cannot provide a new cascading value through the {nameof(SetParametersAndRender)} method.");
+					paramDict.Add(param.Name!, param.Value);
+				}
+				parameterView = ParameterView.FromDictionary(paramDict);
+			}
+			return parameterView;
 		}
 	}
 }