+++ src/bunit.web/TestContext.cs
@@ -32,23 +32,19 @@ namespace Bunit
 			return new RenderedComponent<TComponent>(Services, renderResult.ComponentId, renderResult.Component);
 		}
 
-
 		/// <summary>
 		/// Instantiates and performs a first render of a component of type <typeparamref name="TComponent"/>.
 		/// </summary>
 		/// <typeparam name="TComponent">Type of the component to render</typeparam>
-		/// <param name="componentParameterBuilderAction">The ComponentParameterBuilder action to add type safe parameters to pass to the component when it is rendered</param>
+		/// <param name="parameterBuilder">The ComponentParameterBuilder action to add type safe parameters to pass to the component when it is rendered</param>
 		/// <returns>The rendered <typeparamref name="TComponent"/></returns>
-		public virtual IRenderedComponent<TComponent> RenderComponent<TComponent>(Action<ComponentParameterBuilder<TComponent>> componentParameterBuilderAction) where TComponent : class, IComponent
+		public virtual IRenderedComponent<TComponent> RenderComponent<TComponent>(Action<ComponentParameterBuilder<TComponent>> parameterBuilder) where TComponent : class, IComponent
 		{
-			if (componentParameterBuilderAction is null)
-			{
-				throw new ArgumentNullException(nameof(componentParameterBuilderAction));
-			}
+			if (parameterBuilder is null)
+				throw new ArgumentNullException(nameof(parameterBuilder));
 
 			var builder = new ComponentParameterBuilder<TComponent>();
-			componentParameterBuilderAction(builder);
-
+			parameterBuilder(builder);
 			var renderResult = Renderer.RenderComponent<TComponent>(builder.Build());
 			return new RenderedComponent<TComponent>(Services, renderResult.ComponentId, renderResult.Component);
 		}